Start- og sluttdato er nå påkrevet. Renamet pamelding-ekeltplass.ts -> pamelding-enkeltplass.ts
