+++ apps/nav-veileders-flate/src/api/data/enkeltplass-pamelding.ts
@@ -3,8 +3,8 @@ import { z } from 'zod'
 export const enkeltplassPameldingSchema = z.object({
   beskrivelse: z.string(),
   prisinformasjon: z.string(),
-  startdato: z.string().optional(),
-  sluttdato: z.string().optional(),
+  startdato: z.string(),
+  sluttdato: z.string(),
   arrangorUnderenhet: z.string()
 })
