+++ apps/nav-veileders-flate/src/components/pamelding/handlinger/meld-pa-direkte/MeldPaDirekteModalEnkeltPlass.tsx
@@ -17,7 +17,7 @@ import { getDeltakerNavn } from '../../../../utils/displayText.ts'
 import {
   formToEnkeltplassRequest,
   generateEnkeltplassPameldingRequest
-} from '../../../../utils/pamelding-ekeltplass.ts'
+} from '../../../../utils/pamelding-enkeltplass.ts'
 import { ConfirmInfoCard } from '../../../ConfirmInfoCard.tsx'
 import { useDeltakerContext } from '../../../tiltak/DeltakerContext.tsx'
