+++ apps/nav-veileders-flate/src/components/rediger-pamelding/LagreUtkastButton.tsx
@@ -8,7 +8,7 @@ import { DeltakerResponse } from '../../api/data/deltaker.ts'
 import { useAppContext } from '../../AppContext.tsx'
 import { PameldingEnkeltplassFormValues } from '../../model/PameldingEnkeltplassFormValues.ts'
 import { PameldingFormValues } from '../../model/PameldingFormValues.ts'
-import { formToEnkeltplassRequest } from '../../utils/pamelding-ekeltplass.ts'
+import { formToEnkeltplassRequest } from '../../utils/pamelding-enkeltplass.ts'
 import { generatePameldingRequestFromForm } from '../../utils/pamelding-form-utils.ts'
 import { usePameldingFormContext } from '../pamelding/PameldingFormContext.tsx'
 import { useDeltakerContext } from '../tiltak/DeltakerContext.tsx'
