+++ apps/nav-veileders-flate/src/components/rediger-pamelding/Utkast.tsx
@@ -2,7 +2,7 @@ import { PencilIcon, XMarkIcon } from '@navikt/aksel-icons'
 import { Button } from '@navikt/ds-react'
 import { DeltakerStatusType, Tiltakskode } from 'deltaker-flate-common'
 import { useState } from 'react'
-import { erEnkeltPlass } from '../../utils/pamelding-ekeltplass.ts'
+import { erEnkeltPlass } from '../../utils/pamelding-enkeltplass.ts'
 import { HorisontalLine } from '../HorisontalLine.tsx'
 import { PameldingEnkeltplassForm } from '../pamelding/enkeltplass/PameldingEnkeltplassForm.tsx'
 import { MeldPaDirekteButton } from '../pamelding/handlinger/meld-pa-direkte/MeldPaDirekteButton.tsx'
